Init food

+++ include/wall.hpp
@@ -1,11 +1,9 @@
 #pragma once
 
 #include <iostream>
-#include <vector>
 #include <SFML/Graphics.hpp>
-#include <algorithm>
 
-#include "pvector.hpp"
+#include "constante.hpp"
 
 
 class Wall {
@@ -17,13 +15,13 @@ class Wall {
         ~Wall();
 
         // Getter.
-        int getWall(const int x, const int y) {return walls[y][x];}
+        uint8_t getWall(const int x, const int y) {return walls[y][x];}
 
 
         void addWall(int x, int y);
         void draw(sf::RenderWindow *window);
 
     private:
-        int walls[(int) SIZE_H / SIZE_WALL][(int) SIZE_W / SIZE_WALL];
+        uint8_t walls[(int) SIZE_H / SIZE_WALL][(int) SIZE_W / SIZE_WALL];
 
 };
